Gmsh: - Physical Nodes and Lines get handled as named nodes or elements dependending   on the dimension.
Documentation:
- Title page reflects move from Klagenfurt to Vienna.
- Documented build for Windows .

Windows support:
- Date.bat for uniform dates on Windows
- Read user Id from registry

- Automatic detection of Fortran/C interface using FortranCInterface from
  CMake >= 2.8.7. This means we need at least CMake 2.8.8.

Compiler/Toolchain support:

- Automatic nightly testing of FeSpace at TU Wien has been migrated to
  /dev/null on 15th of December 2012, where instead it should have been
  migrated to a more stable server (wiki). R.I.P

- Manual testing so far...
  + Linux: CentOS 6 and Ubuntu 12.04 (also MinGW cross-compilers)
  + Windows: WinXP and Win7 64-bit.
  + MacOS X: Snow Leopard (10.6) and Lion (10.7)

- It has to be noted that NOT ALL packages from CFSDEPS have been ported
  to all available platforms. This is especially true for auto tools-based
  (configure/make/make install) projects (GMP, MPFR, LIS, etc), since these
  require a UNIX-like shell and the usual UNIX tools (sed, tr, awk...).

C/CXX/Fortran  |  Build Platform  |  Target Platform    | Comments
==============================================================================
GNU/GFortran   |  Linux 32/64     |  Linux 32/64        | versions >= 4.0
------------------------------------------------------------------------------
GNU/IFORT      |  Linux 32/64     |  Linux 32/64        |
------------------------------------------------------------------------------
ICC/GFortran   |  Linux 32/64     |  Linux 32/64        |
------------------------------------------------------------------------------
Intel/IFORT    |  Linux 32/64     |  Linux 32/64        | ICC/IFORT 12.1
------------------------------------------------------------------------------
OPEN64/OPEN64  |  Linux 32/64     |  Linux 32/64        | AMD, open64.net (1)
------------------------------------------------------------------------------
MinGW/MinGW    |  Linux 32/64     |  Win 64 >= Vista    | (2)
------------------------------------------------------------------------------
MinGW/MinGW    |  Win 32/64 >= XP |  Win 32/64 >= XP    | (3)
------------------------------------------------------------------------------
MSVC/IFORT     |  Win 32/64 >= XP |  Win 32/64 >= XP    | IFORT 10.1 (4)
------------------------------------------------------------------------------
Intel/IFORT    |  Win 32/64 >= XP |  Win 32/64 >= XP    | IFORT 10.1 (4)
------------------------------------------------------------------------------
GNU/GFortran   |  MacOS X 32/64   |  MacOS X 32/64      | (5)
------------------------------------------------------------------------------

(1) Open64 Compiler Suite:
    http://developer.amd.com/tools/cpu-development/x86-open64-compiler-suite/
    http://www.open64.net/

(2) MinGW cross-compiler toolchains have been tested under CentOS 6 and
    Ubuntu 12.04 LTS (cf. Developer's Manual). There is no support for OpenMP
    available in the Ubuntu toolchain! The binaries built with these toolchains
    only seem to run on Windows versions equal or newer than Vista.

(3) The 64-bit MinGW GNU compiler TDM-GCC toolchain (4.7.1) from
    http://tdm-gcc.tdragon.net has been tested. Another toolchain is available
    from http://www.equation.com.

(4) Only the build of CFSDEPS has been tested using MSVC 2008 and Intel 10.1
    32-bit compilers.

(5) Since the XCode (https://developer.apple.com/xcode) framework does not
    include a Fortran compiler, the GFortran compilers
    from http://r.research.att.com/tools corresponding to the Xcode versions
    have been used.

Solver Support:

Solver         |  Type       |  openMP parallel | Comments
==============================================================================
UMFPACK        |  direct     |      no (6)      | (7)
------------------------------------------------------------------------------
LIS            |  iterative  |       yes        | (8)
------------------------------------------------------------------------------
ILUPACK        |  iterative  |       no         |
------------------------------------------------------------------------------
SuperLU        |  direct     |       yes        | (9)
------------------------------------------------------------------------------

(6) UMFPACK uses BLAS and LAPACK, therefore parallelization may be available
    through these packages.

(7) Some fields of the UMFPACK Control array have not yet been made available
    in the XML files.

(8) The original implementation has been done by the hero of work Andi Hüppe.
    I just added support for complex matrices.

(9) SuperLU has been integrated into CFSDEPS but not yet into CFS++ itself.
    Just the serial version of SuperLU contains complex ILUs. Therefore, only
    only this version has been integrated into CFSDEPS.

Matrix/Vector IO:

- Full support for MatrixMarket and Harwell-Boeing output for single and SBM
  matrices and vectors.

- MatrixMarket output of the matrix A and the RHS b is done in such a way,
  that one can solve the system Ax = b in Octave/Matlab in the following way:

  > addpath('/CFS_SOURCE_DIR/share/matlab')
  > A = mmread('simname_0_0.mtx');
  > b = mmread('simname_rhs_0.mtx');
  > x = A\b;

- MatrixMarket output is now the default for matrices and vectors. No plain
  ASCII output of the vectors is available anymore.

- MatrixMarket and Harwell-Boeing output format files have been checked with
  the BeBOP sparse matrix conversion tool (http://bebop.cs.berkeley.edu/smc/).

- Added additional Matlab scripts for matrix handling to
  CFS_SOURCE_DIR/share/matlab.

+++ source/DataInOut/SimInOut/gmv/gmvread.cc
@@ -358,7 +358,7 @@ int ignore_me;
 
     for(it=gmv_mem_alloc_set.begin(), eit=gmv_mem_alloc_set.end(); it != eit; it++)
     {
-      LOG_DBG2(gmvread) << "Cleaning up: " << ((unsigned long)*it);
+      LOG_DBG2(gmvread) << "Cleaning up: " << ((size_t)*it);
       free(*it);
     }
     gmv_mem_alloc_set.clear();
